flask-project-struct2x/requirements.txt
alembic==1.11.1
blinker==1.6.2
click==8.1.4
dominate==2.8.0
Flask==2.3.2
Flask-Bootstrap4==4.0.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
greenlet==2.0.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
msgpack==1.0.5
packaging==23.1
pynvim==0.4.3
PyYAML==6.0
Send2Trash==1.8.2
SQLAlchemy==2.0.18
typing_extensions==4.7.1
visitor==0.1.3
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.16.0
